pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectrune"
version = "0.1.0"
description = "Local Home Automation Server for Apple TVs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    "pyatv>=0.14.5",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.1",
    "samsungtvws>=2.7.2",
    "py-samsungtv>=0.7.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["home_automation_server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

